# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "NyanPass 高性能中转面板 API 文档",
        "title": "NyanPass API",
        "termsOfService": "https://nyanpass.io/terms",
        "contact": {
            "name": "API Support",
            "url": "https://nyanpass.io/support",
            "email": "support@nyanpass.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "/api/v1",
    "paths": {
...
        "/api/v1/admin/plans": {
            "get": {
                "tags": [
                    "Admin/Plan"
                ],
                "summary": "获取所有套餐列表（包含隐藏）",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Admin/Plan"
                ],
                "summary": "创建新套餐",
                "parameters": [
                    {
                        "description": "套餐信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.CreatePlanRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/admin/plans/{id}": {
            "get": {
                "tags": [
                    "Admin/Plan"
                ],
                "summary": "获取套餐详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "套餐ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Admin/Plan"
                ],
                "summary": "更新套餐",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "套餐ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "套餐信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.UpdatePlanRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Admin/Plan"
                ],
                "summary": "删除套餐",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "套餐ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/admin/settings": {
            "get": {
                "tags": [
                    "Admin/Setting"
                ],
                "summary": "获取所有设置（管理员）",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Admin/Setting"
                ],
                "summary": "设置单个值（管理员）",
                "parameters": [
                    {
                        "description": "设置信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.SetRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/admin/settings/batch": {
            "put": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Admin/Setting"
                ],
                "summary": "批量更新设置（管理员）",
                "parameters": [
                    {
                        "description": "设置列表",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.BatchUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/admin/settings/group/{group}": {
            "get": {
                "tags": [
                    "Admin/Setting"
                ],
                "summary": "获取指定分组的设置（管理员）",
                "parameters": [
                    {
                        "type": "string",
                        "description": "分组名称",
                        "name": "group",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/admin/settings/{key}": {
            "delete": {
                "tags": [
                    "Admin/Setting"
                ],
                "summary": "删除设置（管理员）",
                "parameters": [
                    {
                        "type": "string",
                        "description": "设置键名",
                        "name": "key",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
        "/client/subscribe/{token}": {
            "get": {
                "description": "根据用户 UUID Token 获取订阅链接，支持多种客户端格式",
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "Subscribe"
                ],
                "summary": "获取订阅配置",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户订阅 Token (UUID)",
                        "name": "token",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "订阅配置内容",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "缺少 Token",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "用户不存在",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "handler.SetRequest": {
            "type": "object",
            "required": [
                "key"
            ],
            "properties": {
                "key": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
...
        "response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "业务码",
                    "type": "integer"
                },
                "data": {
                    "description": "数据"
                },
                "msg": {
                    "description": "提示信息",
                    "type": "string"
                }
            }
        },
...
        "service.BatchUpdateRequest": {
            "type": "object",
            "required": [
                "settings"
            ],
            "properties": {
                "settings": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/service.SettingItem"
                    }
                }
            }
        },
...
        "service.CreatePlanRequest": {
            "type": "object",
            "required": [
                "duration",
                "name"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "device_limit": {
                    "description": "设备数限制",
                    "type": "integer",
                    "minimum": 0
                },
                "duration": {
                    "description": "有效期(天)",
                    "type": "integer"
                },
                "group_id": {
                    "description": "用户组",
                    "type": "integer"
                },
                "hidden": {
                    "description": "是否隐藏",
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "number",
                    "minimum": 0
                },
                "sort": {
                    "description": "排序权重",
                    "type": "integer"
                },
                "speed_limit": {
                    "description": "速度限制(Mbps)",
                    "type": "integer",
                    "minimum": 0
                },
                "transfer": {
                    "description": "流量限制(GB)",
                    "type": "integer",
                    "minimum": 0
                }
            }
        },
...
        "service.SettingItem": {
            "type": "object",
            "required": [
                "key"
            ],
            "properties": {
                "desc": {
                    "type": "string"
                },
                "group": {
                    "type": "string"
                },
                "key": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
...
        "service.UpdatePlanRequest": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "device_limit": {
                    "type": "integer",
                    "minimum": 0
                },
                "duration": {
                    "type": "integer"
                },
                "group_id": {
                    "type": "integer"
                },
                "hidden": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "number",
                    "minimum": 0
                },
                "sort": {
                    "type": "integer"
                },
                "speed_limit": {
                    "type": "integer",
                    "minimum": 0
                },
                "transfer": {
                    "type": "integer",
                    "minimum": 0
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT Token 认证，格式：Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}